requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyJWT==2.9.0
tensorflow==2.19.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==11.0.0
cloudinary==1.36.0